feature(MediaCRUD): new service
